Define the Jacobi symbol and prove the Legendre symbol is periodic (#799)
Figured I would add the Jacobi symbol since the Legendre symbol is now
defined.

I wasn't sure which "squareness" page to link to when mentioning it in
the "Idea" section of the Jacobi symbol, so I just decided to link to
the squareness in `Zp` page since that's the one most relevant for the
Legendre (and herce, Jacobi) symbol. But perhaps there should be no link
there at all.

Here there's also some minor reorganization of the Legendre symbol
definition and a proof of its periodicity (in its top argument).

Also in here are some very minor formatting tweaks to some other
entries. Do let me know if those kinds of changes should be in separate
pull requests from more contentful ones.

---------

Co-authored-by: mrh <mrh57@posteo.net>
Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>