Modal type theory (#701)
### Additions
- Module on modal type theory, including
  - Define the flat modality
  - Define the sharp modality
  - Start work on their adjunction
- Some basic definitions about codiscrete types and the flat modal
operator
- Some additions on higher modalities
- Introduce (strong) subuniverse induction as a relaxation of modal
subuniverse induction
- Show we can induct on the identity types with a higher modality
without appealing to univalence using strong subuniverse induction
  - Start on functorial properties of higher modalities
- Multivariable sections is a definition of sections of multivariable
maps that make it possible to circumvent applying function
extensionality in many cases.
  
(If we at some point in the future want a module on simplicial homotopy
type theory, then it would be fruitful to develop the module on modal
type theory further :))
 
### A regression
It has become clear that the current formulation of higher modalities is
not the most practical one, as it needs to assume everything is
sufficiently contained within a single universe. In addition, the small
relaxation to locally small types seems somewhat contentless and makes
the formalizations more cumbersome. I have started a new line of
formalization to fix these problems in #890.